% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovidVaccinePrediction.R
\docType{package}
\name{CovidVaccinePrediction}
\alias{CovidVaccinePrediction}
\title{CovidVaccinePrediction}
\description{
A package for running prediction model development studies using data in the OMOP CDM
}
