% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{execute}
\alias{execute}
\title{Execute the Study}
\usage{
execute(
  connectionDetails,
  cdmDatabaseSchemas,
  cdmDatabaseNames = "friendly database name",
  cohortDatabaseSchemas = cdmDatabaseSchema,
  cohortTables = "cohort",
  oracleTempSchema = cohortDatabaseSchema,
  outputFolder,
  createCohorts = F,
  fetchData = F,
  runDevelopment = F,
  runValidate = F,
  sampleSize = NULL,
  verbosity = "INFO",
  cdmVersion = 5
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchemas}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cdmDatabaseNames}{Shareable name of the database}

\item{cohortDatabaseSchemas}{Schema name where intermediate data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{cohortTables}{The name of the table that will be created in the work database schema.
This table will hold the target population cohorts used in this
study.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{createCohorts}{Create the cohortTable table with the target population and outcome cohorts?}

\item{fetchData}{Only fetch data for the analyses without fitting models. Setting this flag will overwrite your input provided to the runAnalyses and createCohorts parameters.}

\item{runDevelopment}{Run the model development}

\item{runValidate}{Run the model validation}

\item{sampleSize}{The number of patients in the target cohort to sample (if NULL uses all patients)}

\item{verbosity}{Sets the level of the verbosity. If the log level is at or higher in priority than the logger threshold, a message will print. The levels are:
\itemize{
\item{DEBUG}{Highest verbosity showing all debug statements}
\item{TRACE}{Showing information about start and end of steps}
\item{INFO}{Show informative information (Default)}
\item{WARN}{Show warning messages}
\item{ERROR}{Show error messages}
\item{FATAL}{Be silent except for fatal errors}
}}

\item{cdmVersion}{The version of the common data model}

\item{createProtocol}{Creates a protocol based on the analyses specification}
}
\description{
Execute the Study
}
\details{
This function executes the CovidVaccinePrediction Study.
}
\examples{
\dontrun{
connectionDetails <- createConnectionDetails(dbms = "postgresql",
                                             user = "joe",
                                             password = "secret",
                                             server = "myserver")

execute(connectionDetails,
        cdmDatabaseSchemas = "cdm_data",
        cdmDatabaseNames = 'shareable name of the database'
        cohortDatabaseSchemas = "study_results",
        cohortTables = "cohort",
        oracleTempSchema = NULL,
        outputFolder = "c:/temp/study_results", 
        createCohorts = T,
        fetchData = T,
        runDevelopment = T,
        sampleSize = 10000,
        verbosity = "INFO",
        cdmVersion = 5)
}

}
